+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2] - 2026-09-20
+
+### Changed
+
+- When a PR is opened for release, the semver max is always used. If the PR 
+  already exists and the version is different, then the old PR is closed and a
+  new PR is created.
+
 ## [2.0.1] - 2026-09-05
 
 ### Fixed